services/redis-aggregator/requirements.txt
confluent-kafka==2.3.0
redis==5.0.1
pyyaml==6.0.1